requirements/base.txt
# Core
Django==3.2.1
djangorestframework==3.12.4

# Downloading
requests==2.25.1

# AWS
boto3==1.17.65
